// src/api/BidragBeregnForskuddApi.ts
export interface Grunnlag {
    referanse?: string;
    type?: string;
    innhold?: JsonNode;
}

export type JsonNode = object;